Refactor build system and update project structure
Updated .clang-tidy configuration to change HeaderFilterRegex and remove a check. Updated CMakeLists.txt to set C++20 standard, change paths, add knoodle-engine target, and update documentation paths. Added CMakePresets.json for build configurations. Updated Doxyfile.in to change INPUT directory. Updated KnoodleModule.cmake for warning suppression and include directories. Removed core-config.cmake.in and ghi-config.cmake.in. Added knoodle-engine-config.cmake. Updated common.hpp to include knoodle_engine_api.h. Refactored include paths in various source files. Added os_linux.cpp for shared library handling on Linux. Removed knoodle_create_module calls for core and ghi modules.
