+++ cmake/KnoodleModule.cmake
@@ -26,14 +26,14 @@ function(knoodle_create_module)
 
   target_compile_options(${KNOODLE_MODULE_NAME}
     PRIVATE
-      $<$<CXX_COMPILER_ID:MSVC>:/W4 /WX /wd4251>
+      $<$<CXX_COMPILER_ID:MSVC>:/W4 /WX /wd4251 /wd4275>
       $<$<NOT:$<CXX_COMPILER_ID:MSVC>>:-Wall -Wextra -pedantic -Werror>)
 
   # Set the module properties
   target_include_directories(${KNOODLE_MODULE_NAME}
     PUBLIC
-      $<BUILD_INTERFACE:${KNOODLE_ROOT_DIR}/include/${KNOODLE_MODULE_NAME}>
-      $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/generated>
+      $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include/${KNOODLE_MODULE_NAME}>
+      $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/generated>
 
       $<INSTALL_INTERFACE:include>)
 
@@ -54,25 +54,6 @@ function(knoodle_create_module)
     PREFIX_NAME KN_
     EXPORT_MACRO_NAME ${MODULE_API_NAME})
 
-  # Setup module installation
-  install(TARGETS ${KNOODLE_MODULE_NAME}
-        RUNTIME           # Following options apply to runtime artifacts.
-          COMPONENT Runtime
-        LIBRARY           # Following options apply to library artifacts.
-          COMPONENT Runtime
-          NAMELINK_COMPONENT Development
-        ARCHIVE           # Following options apply to archive artifacts.
-          COMPONENT Development
-          DESTINATION lib/static
-        FILE_SET HEADERS  # Following options apply to file set HEADERS.
-          COMPONENT Development)
-
-  install(DIRECTORY "${KNOODLE_ROOT_DIR}/include/${KNOODLE_MODULE_NAME}"
-    DESTINATION include)
-
-  install(DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}/generated/"
-    DESTINATION include/${KNOODLE_MODULE_NAME})
-
   # force include the export header
   target_compile_options(${KNOODLE_MODULE_NAME}
     PUBLIC
@@ -81,14 +62,29 @@ function(knoodle_create_module)
 
   include(CMakePackageConfigHelpers)
   write_basic_package_version_file(
-    "${CMAKE_CURRENT_BINARY_DIR}/${KNOODLE_MODULE_NAME}ConfigVersion.cmake"
+    "${CMAKE_BINARY_DIR}/${KNOODLE_MODULE_NAME}ConfigVersion.cmake"
     VERSION ${PROJECT_VERSION}
     COMPATIBILITY AnyNewerVersion)
 
   configure_package_config_file(
-    "${KNOODLE_ROOT_DIR}/cmake/${KNOODLE_MODULE_NAME}-config.cmake.in"
-    "${CMAKE_CURRENT_BINARY_DIR}/${KNOODLE_MODULE_NAME}-config.cmake"
+    "${PROJECT_SOURCE_DIR}/cmake/${KNOODLE_MODULE_NAME}-config.cmake.in"
+    "${CMAKE_BINARY_DIR}/${KNOODLE_MODULE_NAME}-config.cmake"
     INSTALL_DESTINATION "${CMAKE_INSTALL_LIBDIR}/cmake/${KNOODLE_MODULE_NAME}")
+
+# Setup module installation
+install(TARGETS ${KNOODLE_MODULE_NAME}
+      RUNTIME           # Following options apply to runtime artifacts.
+        COMPONENT Runtime
+      LIBRARY           # Following options apply to library artifacts.
+        COMPONENT Runtime
+        NAMELINK_COMPONENT Development
+      ARCHIVE           # Following options apply to archive artifacts.
+        COMPONENT Development
+        DESTINATION lib/static
+      FILE_SET HEADERS  # Following options apply to file set HEADERS.
+        COMPONENT Development)
+
+install(DIRECTORY "${PROJECT_SOURCE_DIR}/source" DESTINATION include FILES_MATCHING PATTERN "*.hpp")
 
 endfunction()
