+++ source/core/kn_assert.hpp
