+++ source/core/string_utils.hpp
