+++ source/ghi/ghi_definitions.hpp
