+++ source/knoodle_version.hpp
